refactor(storage): extract _MarketDataMixin
Moves 10 market-data methods (upsert_daily_ohlc, list_daily_ohlc,
upsert_intraday_quote, get_intraday_quote, get_latest_intraday_quote_times,
append_pcr_history, get_pcr_history_30d_ago, get_pcr_history_row,
get_recent_etf_aum, upsert_etf_aum) to storage/market_data.py.
