+++ src/uw_scan/storage/repository.py
@@ -38,6 +38,7 @@ from ._helpers import (
     status_family_for,
 )
 from .audit import _AuditMixin
+from .market_data import _MarketDataMixin
 
 # Row dataclasses live in rows.py since the PR-1 split. Re-exported here so
 # existing callers (`from uw_scan.storage.repository import JobRow`) continue
@@ -331,7 +332,7 @@ _RECORD_HEALTH_EXCLUDED_TABLES = {
 logger = logging.getLogger(__name__)
 
 
-class Repository(_AuditMixin, _ScanOutputsMixin, _BaseMixin):
+class Repository(_AuditMixin, _MarketDataMixin, _ScanOutputsMixin, _BaseMixin):
     """Repository wraps a psycopg connection and exposes typed CRUD.
 
     Inherits __init__ and the conn property from _BaseMixin. As PR-1/2/3
@@ -3981,125 +3982,7 @@ class Repository(_AuditMixin, _ScanOutputsMixin, _BaseMixin):
         ]
         return cards, summary
 
-    # ---- daily_ohlc ----
-    def upsert_daily_ohlc(
-        self,
-        *,
-        ticker: str,
-        date: _date,
-        open: Decimal | None,
-        high: Decimal | None,
-        low: Decimal | None,
-        close: Decimal,
-        volume: int | None,
-        source: str,
-    ) -> None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                INSERT INTO {self._schema}.daily_ohlc
-                  (ticker, date, open, high, low, close, volume, source)
-                VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
-                ON CONFLICT (ticker, date) DO UPDATE
-                  SET open=EXCLUDED.open, high=EXCLUDED.high, low=EXCLUDED.low,
-                      close=EXCLUDED.close, volume=EXCLUDED.volume,
-                      source=EXCLUDED.source, fetched_at=NOW()
-                """,
-                (ticker, date, open, high, low, close, volume, source),
-            )
-        self._conn.commit()
-
-    def list_daily_ohlc(self, ticker: str, *, limit: int = 30) -> list[DailyOhlcRow]:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT ticker, date, open, high, low, close, volume, source, fetched_at
-                FROM {self._schema}.daily_ohlc
-                WHERE ticker=%s
-                ORDER BY date DESC
-                LIMIT %s
-                """,
-                (ticker, limit),
-            )
-            return [DailyOhlcRow(*row) for row in cur.fetchall()]
-
-    # ---- intraday_quote ----
-    def upsert_intraday_quote(
-        self, ticker: str, price: Decimal, quoted_at: datetime
-    ) -> None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                INSERT INTO {self._schema}.intraday_quote (ticker, price, quoted_at)
-                VALUES (%s, %s, %s)
-                ON CONFLICT (ticker) DO UPDATE
-                  SET price=EXCLUDED.price, quoted_at=EXCLUDED.quoted_at, fetched_at=NOW()
-                """,
-                (ticker, price, quoted_at),
-            )
-        self._conn.commit()
-
-    def get_intraday_quote(self, ticker: str) -> IntradayQuoteRow | None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT ticker, price, quoted_at, fetched_at
-                FROM {self._schema}.intraday_quote WHERE ticker=%s
-                """,
-                (ticker,),
-            )
-            row = cur.fetchone()
-            return IntradayQuoteRow(*row) if row else None
-
-    def get_latest_intraday_quote_times(self) -> tuple[datetime, datetime] | None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT MAX(quoted_at), MAX(fetched_at)
-                FROM {self._schema}.intraday_quote
-                """
-            )
-            row = cur.fetchone()
-        if row and row[0] is not None and row[1] is not None:
-            return (row[0], row[1])
-        return None
-
-    # ---- pcr_history ----
-    def append_pcr_history(
-        self,
-        ticker: str,
-        snapshot_date: _date,
-        pcr_oi: Decimal | None,
-        pcr_vol: Decimal | None,
-    ) -> None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                INSERT INTO {self._schema}.pcr_history (ticker, snapshot_date, pcr_oi, pcr_vol)
-                VALUES (%s, %s, %s, %s)
-                ON CONFLICT (ticker, snapshot_date) DO UPDATE
-                  SET pcr_oi=EXCLUDED.pcr_oi, pcr_vol=EXCLUDED.pcr_vol
-                """,
-                (ticker, snapshot_date, pcr_oi, pcr_vol),
-            )
-        self._conn.commit()
-
-    def get_pcr_history_30d_ago(
-        self, ticker: str, today: _date
-    ) -> PcrHistoryRow | None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT ticker, snapshot_date, pcr_oi, pcr_vol
-                FROM {self._schema}.pcr_history
-                WHERE ticker=%s AND snapshot_date <= %s - INTERVAL '30 days'
-                ORDER BY snapshot_date DESC
-                LIMIT 1
-                """,
-                (ticker, today),
-            )
-            row = cur.fetchone()
-            return PcrHistoryRow(*row) if row else None
+    # daily_ohlc / intraday_quote / pcr_history methods moved to _MarketDataMixin
 
     # ---- jobs ----
     def enqueue_rescan_job(self, ticker: str, *, priority: int = 0) -> str:
@@ -4232,35 +4115,7 @@ class Repository(_AuditMixin, _ScanOutputsMixin, _BaseMixin):
             oldest_requested_at=row[3],
         )
 
-    # ---- etf_aum_cache (A1 review fix: skip per-scan UW round trip) ----
-    def get_recent_etf_aum(self, ticker: str, *, max_age: timedelta) -> Decimal | None:
-        """Return cached AUM if fetched within max_age, else None.
-        None means the caller should fetch fresh (cache miss or stale)."""
-        ticker = ticker.upper()  # cache keys are canonical UPPER
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT aum FROM {self._schema}.etf_aum_cache
-                WHERE ticker = %s AND fetched_at > NOW() - %s
-                """,
-                (ticker, max_age),
-            )
-            row = cur.fetchone()
-        return row[0] if row else None
-
-    def upsert_etf_aum(self, ticker: str, aum: Decimal) -> None:
-        ticker = ticker.upper()
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                INSERT INTO {self._schema}.etf_aum_cache (ticker, aum, fetched_at)
-                VALUES (%s, %s, NOW())
-                ON CONFLICT (ticker) DO UPDATE
-                  SET aum = EXCLUDED.aum, fetched_at = EXCLUDED.fetched_at
-                """,
-                (ticker, aum),
-            )
-        self._conn.commit()
+    # etf_aum_cache methods moved to _MarketDataMixin
 
     # ---- aggregates (JSONB on scan_runs) ----
     def set_aggregates(self, run_id: int, agg: "models.MarketAggregates") -> None:
@@ -4282,20 +4137,7 @@ class Repository(_AuditMixin, _ScanOutputsMixin, _BaseMixin):
             return None
         return models.MarketAggregates.model_validate(row[0])
 
-    def get_pcr_history_row(
-        self, ticker: str, snapshot_date: _date
-    ) -> PcrHistoryRow | None:
-        with self._conn.cursor() as cur:
-            cur.execute(
-                f"""
-                SELECT ticker, snapshot_date, pcr_oi, pcr_vol
-                FROM {self._schema}.pcr_history
-                WHERE ticker=%s AND snapshot_date=%s
-                """,
-                (ticker, snapshot_date),
-            )
-            row = cur.fetchone()
-        return PcrHistoryRow(*row) if row else None
+    # get_pcr_history_row moved to _MarketDataMixin
 
     # ---- stock history rollup (for Market Structure history table) ----
     def fetch_stock_history_rollup(self, ticker: str, limit: int = 30) -> list[dict]:
